scoreboard/prototype/serializers.py
from rest_framework import serializers
from .models import TaskGroup, Task, UserTask


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ["id", "name"]


class TaskGroupSerializer(serializers.ModelSerializer):
    tasks = serializers.StringRelatedField(many=True)

    class Meta:
        model = TaskGroup
        fields = ["id", "name", "tasks"]


class UserTaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserTask
        fields = "__all__"
