pyproject.toml
[tool.poetry]
name = "scoreboard"
version = "0.1.0"
description = ""
authors = ["TheBlackFern <doroginin.dv@phystech,edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
pylint-django = "^2.5.3"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
